refactoring initForm
